\name{barley}

\alias{barley}

\docType{data}

\title{
A barley genotype dataset
}

\description{
A data frame of barley genotypes with genetic distance
}

\usage{data(barley)}
\format{
  A data frame with 100 observations on the following 36 variables(barley SNP markers).
  \describe{The first row is the genetic distance.}
}


\examples{
data(barley)

}

\keyword{datasets}
