\name{linkim-package}
\alias{linkim-package}
\alias{linkim}
\docType{package}
\title{
Linkage information based genotype imputation method
}
\description{
A linkage information based method for imputing missing diploid genotypes
}

\details{
\tabular{ll}{
Package: \tab linkim\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-14\cr
License: GPL(>= 2)\cr
}
}

\author{
Yi Xu and Jixiang Wu\cr
Maintainer: Yi Xu <statxy@gmail.com>
}


\keyword{package}


